move podtemplatespec to package mutators, use empty struct instead
