+++ internal/mutators/constants.go
@@ -1,0 +1,5 @@
+package mutators
+
+const (
+	defaultLimitRequestMemoryRatio float64 = 1.1
+)
